+++ recon/symbolized/app/__ieee754_exp.c
@@ -19,26 +19,26 @@
  */
 /* Reconstructed FUN_00012db0 @ 0x12db0  (parity: 300/300 trials, PROVEN) */
 #include <stdint.h>
-typedef uint32_t u4; typedef int32_t i4; typedef uint64_t u8; typedef int64_t i8;
+typedef uint32_t u4; typedef int32_t i4; typedef uint64_t u64; typedef int64_t i64;
 extern unsigned long long dcmp_negate_rhs(unsigned int, unsigned int, unsigned int, unsigned int);
-extern u8 __aeabi_dadd(u4,u4,u4,u4);
-extern u8 __floatsidf(void);
-extern u8 __muldf3(u4,u4,u4,u4);
-extern u8 __divdf3(u4,u4,u4,u4);
-extern u8 __aeabi_dcmplt(u4,u4,u4,u4);
-extern u8 __aeabi_dcmpgt(u4,u4,u4,u4);
-extern u8 __fixdfsi(void);
+extern u64 __aeabi_dadd(u4,u4,u4,u4);
+extern u64 __floatsidf(void);
+extern u64 __muldf3(u4,u4,u4,u4);
+extern u64 __divdf3(u4,u4,u4,u4);
+extern u64 __aeabi_dcmplt(u4,u4,u4,u4);
+extern u64 __aeabi_dcmpgt(u4,u4,u4,u4);
+extern u64 __fixdfsi(void);
 
 double __ieee754_exp(double xd, u4 param_1, u4 param_2, u4 param_3, u4 param_4){
-    u8 in_d0; __builtin_memcpy(&in_d0,&xd,8);
-    u8 uVar15 = ((u8)param_2<<32)|param_1;
-    u8 uVar12=0, uVar13=0, uVar14=0x7e37e43c8800759cULL, uVar16=0;
+    u64 in_d0; __builtin_memcpy(&in_d0,&xd,8);
+    u64 uVar15 = ((u64)param_2<<32)|param_1;
+    u64 uVar12=0, uVar13=0, uVar14=0x7e37e43c8800759cULL, uVar16=0;
     u4 uVar9 = (u4)(in_d0>>32);
     u4 uVar5 = uVar9 & 0x7fffffff;
     i4 iVar4 = (i4)uVar9 >> 31;
     i4 iVar1, iVar2=0, iVar7=0;
     u4 uVar3=0, uVar6=0, uVar8, uVar10;
-    u8 res;
+    u64 res;
     const u4 *puVar11;
 
     if (0x40862e41u < uVar5) {
@@ -46,10 +46,10 @@ double __ieee754_exp(double xd, u4 param_1, u4 param_2, u4 param_3, u4 param_4){
             if ((in_d0 & 0xfffff00000000ULL)!=0 || (i4)(u4)in_d0 != 0) {
                 res = __aeabi_dadd((u4)in_d0,uVar9,(u4)in_d0,uVar9); goto ret;
             }
-            if ((i8)in_d0 >= 0) { res = in_d0; goto ret; }
+            if ((i64)in_d0 >= 0) { res = in_d0; goto ret; }
         } else {
             iVar1 = (i4)(u4)__aeabi_dcmpgt((u4)in_d0,uVar9,0xfefa39efu,0x40862e42u);
-            uVar12 = ((u8)param_4<<32)|param_3;
+            uVar12 = ((u64)param_4<<32)|param_3;
             uVar3=0x8800759cu; iVar2=0x7e37e43c; uVar6=0x8800759cu; iVar7=0x7e37e43c;
             if (iVar1 != 0) goto LAB_12e26;
             iVar2 = (i4)(u4)__aeabi_dcmplt((u4)in_d0,uVar9,0xfefa39efu,0x40862e42u);
@@ -78,7 +78,7 @@ LAB_12f2e:
                 puVar11 = (const u4*)(((unsigned long)&rodata_88858) /*=0x88858*/ + (u4)(iVar4 * -8));
                 uVar14 = dcmp_negate_rhs((u4)in_d0,uVar9,
                                       puVar11[0],puVar11[1]);
-                uVar15 = *(const u8*)(((unsigned long)&rodata_88848) /*=0x88848*/ + (u4)(iVar4 * -8));
+                uVar15 = *(const u64*)(((unsigned long)&rodata_88848) /*=0x88848*/ + (u4)(iVar4 * -8));
             }
             in_d0 = dcmp_negate_rhs((u4)uVar14,(u4)(uVar14>>32),(u4)uVar15,(u4)(uVar15>>32));
         } else if (0x3e2fffffu < uVar5) {
@@ -120,7 +120,7 @@ LAB_12f2e:
         uVar14 = dcmp_negate_rhs(0,0x3ff00000u,(u4)uVar14,(u4)(uVar14>>32));
         iVar4 = (i4)(u4)(uVar14>>32);
         if (-0x3fe < iVar2) {
-            res = ((u8)(u4)(iVar4 + iVar2*0x100000)<<32) | (u4)uVar14;
+            res = ((u64)(u4)(iVar4 + iVar2*0x100000)<<32) | (u4)uVar14;
             goto ret;
         }
         uVar3 = (u4)uVar14;
